requirements.txt
mysql-connector==2.2.9
selenium==3.141.0
urllib3==1.26.3
